【WA】[第７回日本情報オリンピック 本選（オンライン）]E - ペンキの色

+++ joi2008ho/joi2008ho_e.cpp
@@ -9,7 +9,7 @@ const double PI = 3.14159265358979323846;
 typedef long long ll;
 const double EPS = 1e-9;
 #define rep(i, n) for (int i = 0; i < (n); ++i)
-typedef pair<unsigned short, unsigned short> P;
+typedef pair<int, int> P;
 const ll INF = 10e17;
 #define cmin(x, y) x = min(x, y)
 #define cmax(x, y) x = max(x, y)
@@ -226,6 +226,8 @@ Conv make_horizontal_conv(vector<Tape> &tapes, ll w) {
             horizon_conv.cache(x);
         }
     }
+
+
     horizon_conv.build();
     return horizon_conv;
 }
